requirements.txt
google-genai>=1.62.0
python-dotenv>=1.0.1
pandas>=2.2.0
